Get rid of index file related warnings.
git-svn-id: https://svn.code.sf.net/p/pymca/code@1503 a8702ac3-0063-49a2-b157-4a214673c9c1

+++ PyMca/specfile/src/sfdata.c
@@ -411,7 +411,7 @@ if(0){
               data[rows] = dataline;
               dinfo[ROW]=rows+1;
         }else{
-              printf("Error on scan %d line %d\n", index, rows+1);
+              printf("Error on scan %d line %d\n", (int) index, (int) (rows+1));
               /* just ignore the line instead of stopping there with a
               break; */
               rows--;
